+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 ### Fixed
 
 - call the elements `focus` method inside of `focusable` to set focus
+- call the elements `blur` method inside of `blurrable` to unset focus
 
 ## [0.7.2] - 2018-07-23